+++ corporal/httphelp/common.go
@@ -5,7 +5,13 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+
+	"github.com/gorilla/mux"
 )
+
+type HandlerRegistrator interface {
+	RegisterRoutesWithRouter(router *mux.Router)
+}
 
 func readBytesAndRecreateReader(source io.ReadCloser) ([]byte, io.ReadCloser, error) {
 	// Reading an unlimited amount of data might be dangerous.